requirement.txt
streamlit
plotly==5.8.0  
pandas
psycopg2-binary
#database collector
python-dotenv